Fix cleanup again
When you unbind the device while in use, the pci core tries to
tear down the IRQs and thus we hit this bug:

kernel BUG at drivers/pci/msi.c:371!

Instead we'll have to drop our IRQs during unregister and
re-introduce an alive parameter. But this time alive will be
protected by the mrpc_mutex and we will be very careful
about how things will be torn down.