switchtec_ntb: initialize hardware for doorbells and messages
This patch sets up some hardware registers and creates interupt service
routines for the doorbells and messages.

There are 64 doorbells in the switch that are shared between all
partitions. The upper 4 doorbells are also shared with the messages
and are there for not used. Thus, this code provides 28 doorbells
for each partition.