switchtec: add link event notifier block
In order for the switchtec NTB code to handle link change events we
create a notifier block in the switchtec code which gets called
whenever an appropriate event interrupt occurs.

In order to preserve userspaces ability to follow these events,
we compare the event count with a stored copy from last time we
checked.