ntb_hw_switchtec: remove unnecessary shared MW reinit in crosslink case
The reinit of shared MW when receive FORCE_LINK_DOWN message is
required for non-crosslink case, because the LUT settings for shared
MW of peer is reset after reboot.
But in crosslink case, the LUT setttings in the virtual partition
will not be impact by the peer's reboot, therefore it is unnecessary
to reinit it.

Signed-off-by: Wesley Sheng <wesley.sheng@microchip.com>