Remove code for disabling ID protection
ID protection is a static setting in the Switchtec config file. With
ID protection enabled, only posted requests with requester IDs in the
requester ID table will be allowed to access the NT windows. Otherwise
all posted requests are allowed.

ID protect is not applicable to non-posted requests.

There's no need to disable it in the driver as we have already configured
the requester IDs. The driver works no matter this setting is enabled or
not.