Error out MRPC execution when no more GAS access
It was observed that after a firmware hard reset, the MRPC execution
will hang indefinitely. This is because after the reset, the host will
get all 1s for any BAR read, in which case the driver won't get a valid
MRPC status.

Add a read check to GAS access when a MRPC command execution don't
response timely, error out if the check fails.