Setup crosslink requester IDs when updating link status
When accessing the peer status for link status update, we need to setup
the crosslink requester IDs first to ensure we have access to the peer
status.

Earlier crosslink requester IDs setup during device initialization or
link enabling can fail because the peer switch may not in a good state
or the link to the peer switch may not be well connected at the moment.

Therefore, we add the requester IDs setup during link status update
which could be triggered by a peer link up event at which the peer
switch and the connection to the peer switch must be in a good state.

For more background about the requester IDs setup, please refer to
commit fd0edeee119ee3684d1e2e111cb6fe09a636dbb9.