Update link status when enabling link
When enabling link, the crosslink may not be properly configured or the
peer switch may not be probed/initialized by the driver. Call
switchtec_ntb_link_status_update() to try to configure it before sending
the link up message to peer. If the peer switch is probed/initialized
when enabling the link, the call to switchtec_ntb_link_status_update()
will make the crosslink configured, then the link up message will be
able to be sent to the peer switch. Otherwise we will need to wait for
the peer driver to come up and bring up the link.