Fix a crosslink target partition configuration issue
The remote crosslink partition needs to have its outbound message mapped
to the remote host partition which connects to the remote NT host, so
that when a message from the local switch hits the remote crosslink
partition, it can be forwarded to the remote host. Program the remote
host partition ID into the corresponding message registers of the remote
crosslink partition.