Push two reserved LUTs setup to link status update
We need to know the partition ID of remote crosslink partition before we
can access the per-partition NT Control and NT Doorbell/Message
regsters. However, the remote switch may not be ready when the local
one is probed, therefore we can't get the partition ID of remote
crosslink partition and setup the reserved LUTs to access those
registers. Push the initialization of the two reserved LUTs to the link
status update code. A new function is introduced for the deferred part
of the crosslink initialization.